Refactor storage in jikan, add template for StorageTest
