+++ .gitignore
@@ -1,3 +1,6 @@
+# data files
+/data/
+
 # IDEA files
 /.idea/
 /out/
