+++ src/main/java/jikan/storage/Storage.java
@@ -1,4 +1,4 @@
-package storage;
+package jikan.storage;
 
 import java.io.File;
 import java.io.FileWriter;
