+++ src/main/java/jikan/storage/StorageHandler.java
@@ -1,4 +1,4 @@
-package storage;
+package jikan.storage;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
