+++ src/test/java/jikan/storage/StorageTest.java
@@ -1,10 +1,8 @@
-package seedu.duke;
-
+package jikan.storage;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class StorageTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
